remember_maps/maps/models.py
from django.contrib.auth import get_user_model
from django.db import models

from treasuremap.fields import LatLongField


User = get_user_model()


class Post(models.Model):
    text = models.TextField('Текст')
    pub_date = models.DateTimeField('Дата публикации', auto_now_add=True)
    author = models.ForeignKey(
        User,
        on_delete=models.CASCADE,
        related_name='posts',
        verbose_name='Автор'
    )
    point = LatLongField()
    lonlatstr = models.CharField('Для запроса', max_length=255)

    class Meta:
        ordering = ('-pub_date',)
        verbose_name = 'Пост'
        verbose_name_plural = 'Посты'
