requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
Django==4.1.4
django-debug-toolbar==3.8.1
django-treasuremap==0.3.4
sqlparse==0.4.3
